+++ src/components/Page Header/index.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import './styles.scss';
+
+class PageHeader extends React.Component {
+  render() {
+    const { children } = this.props;
+
+    return (
+      <div className="row blog-header">
+        <div className="col-md-8">{children}</div>
+      </div>
+    );
+  }
+}
+
+export default PageHeader;