+++ src/scenes/Work/index.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import PageHeader from '../../components/Page Header';
+
+class Work extends React.Component {
+  render() {
+    return (
+      <div className="container">
+        <PageHeader>
+          <h1>Work</h1>
+          <h6>
+            Lorem ipsum dolor sit amet, consectetur adipisicing elit. Eius sequi
+            delectus doloribus repellendus consequatur atque alias nemo dolorum
+            quidem, magni magnam possimus, fugiat error placeat quam itaque quis
+            quisquam ipsa!
+          </h6>
+        </PageHeader>
+      </div>
+    );
+  }
+}
+
+export default Work;